+++ examples/posix/src/my_platform.c
@@ -112,6 +112,8 @@ static void posix_on_controller_data(uni_hid_device_t* d, uni_controller_t* ctl)
             // Axis ry: control rumble
             if ((gp->buttons & BUTTON_A) && d->report_parser.set_dual_rumble != NULL) {
                 d->report_parser.set_dual_rumble(d, 1000, 100, 0, 255);
+            } else if ((gp->buttons & BUTTON_A) && d->report_parser.set_rumble != NULL) {
+                d->report_parser.set_rumble(d, 0x80, 50);
             }
 
             if ((gp->buttons & BUTTON_B) && d->report_parser.set_dual_rumble != NULL) {